requirements.txt
flask==2.3.2
tensorflow==2.12.0
numpy==1.23.5
pillow==9.5.0
opencv-python-headless==4.8.0.76